Properly isolate the ITs in a Failsafe execution
The ITs were run as part of the Surefire execution and apparently
something is leaking from them.
It wasn't an issue before because they were executed at the end of the
run but they were executed first with the new orderer given they are
not annotated with a Quarkus test annotation.
