+++ integration-tests/test-extension/tests/src/test/java/io/quarkus/it/extension/it/CallbackTestModeIT.java
@@ -1,7 +1,5 @@
 package io.quarkus.it.extension.it;
 
-import static org.assertj.core.api.Assertions.assertThat;
-
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
@@ -13,38 +11,50 @@ import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
 
 import io.quarkus.maven.it.RunAndCheckMojoTestBase;
 import io.quarkus.maven.it.continuoustesting.ContinuousTestingMavenTestUtils;
+import io.quarkus.maven.it.continuoustesting.TestModeContinuousTestingMavenTestUtils;
+import io.quarkus.runtime.LaunchMode;
 
 /**
  * Be aware! This test will not run if the name does not start with 'Test'.
  * <p>
- * NOTE to anyone diagnosing failures in this test, to run a single method use:
- * <p>
- * mvn install -Dit.test=TestTemplateDevModeIT#methodName
+ * This tests test templates, in test mode.
  */
-@DisabledIfSystemProperty(named = "quarkus.test.native", matches = "true")
 @Disabled // Tracked by #22611
-public class TestCallbackDevModeIT extends RunAndCheckMojoTestBase {
+@DisabledIfSystemProperty(named = "quarkus.test.native", matches = "true")
+public class CallbackTestModeIT extends RunAndCheckMojoTestBase {
 
     /*
      * We have a few tests that will run in parallel, so set a unique port
      */
     protected int getPort() {
-        return 8094;
+        return 8096;
     }
 
-    protected void runAndCheck(boolean performCompile, String... options)
-            throws MavenInvocationException, FileNotFoundException {
-        run(performCompile, options);
+    @Override
+    protected LaunchMode getDefaultLaunchMode() {
+        return LaunchMode.TEST;
+    }
 
-        try {
-            String resp = devModeClient.getHttpResponse();
-            assertThat(resp).containsIgnoringCase("ready").containsIgnoringCase("application")
-                    .containsIgnoringCase("SNAPSHOT");
-        } catch (Exception e) {
-            e.printStackTrace();
+    @Override
+    public void shutdownTheApp() {
+        if (running != null) {
+            running.stop();
         }
 
-        // There's no json endpoints, so nothing else to check
+        // There's no http server, so there's nothing to check to make sure we're stopped, except by the maven invoker itself, or the logs
+    }
+
+    /**
+     * This is actually more like runAndDoNotCheck, because
+     * we can't really check anything via a HTTP get, because this is a test mode application
+     */
+    @Override
+    protected void runAndCheck(boolean performCompile, LaunchMode mode, String... options)
+            throws FileNotFoundException, MavenInvocationException {
+        run(performCompile, mode, options);
+
+        // We don't need to try and pause, because the continuous testing utils will wait for tests to finish
+
     }
 
     @Test
@@ -54,12 +64,12 @@ public class TestCallbackDevModeIT extends RunAndCheckMojoTestBase {
         testDir = initProject("projects/project-using-test-callback-from-extension", executionDir);
         runAndCheck();
 
-        ContinuousTestingMavenTestUtils testingTestUtils = new ContinuousTestingMavenTestUtils(getPort());
+        ContinuousTestingMavenTestUtils testingTestUtils = new TestModeContinuousTestingMavenTestUtils(running);
         ContinuousTestingMavenTestUtils.TestStatus results = testingTestUtils.waitForNextCompletion();
         // This is a bit brittle when we add tests, but failures are often so catastrophic they're not even reported as failures,
         // so we need to check the pass count explicitly
         Assertions.assertEquals(0, results.getTestsFailed());
-        Assertions.assertEquals(2, results.getTestsPassed());
+        Assertions.assertEquals(9, results.getTestsPassed());
     }
 
 }
