+++ integration-tests/test-extension/tests/src/test/java/io/quarkus/it/extension/it/ParameterDevModeIT.java
@@ -1,0 +1,57 @@
+package io.quarkus.it.extension.it;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+import java.io.FileNotFoundException;
+import java.io.IOException;
+
+import org.apache.maven.shared.invoker.MavenInvocationException;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.DisabledIfSystemProperty;
+
+import io.quarkus.maven.it.RunAndCheckMojoTestBase;
+import io.quarkus.maven.it.continuoustesting.ContinuousTestingMavenTestUtils;
+
+/**
+ * Be aware! This test will not run if the name does not start with 'Test'.
+ * <p>
+ * NOTE to anyone diagnosing failures in this test, to run a single method use:
+ * <p>
+ * mvn install -Dit.test=DevMojoIT#methodName
+ */
+@DisabledIfSystemProperty(named = "quarkus.test.native", matches = "true")
+public class ParameterDevModeIT extends RunAndCheckMojoTestBase {
+
+    protected int getPort() {
+        return 8098;
+    }
+
+    protected void runAndCheck(boolean performCompile, String... options)
+            throws MavenInvocationException, FileNotFoundException {
+        run(performCompile, options);
+
+        String resp = devModeClient.getHttpResponse();
+
+        assertThat(resp).containsIgnoringCase("ready").containsIgnoringCase("application")
+                .containsIgnoringCase("1.0-SNAPSHOT");
+
+        // There's no json endpoints, so nothing else to check
+    }
+
+    @Test
+    public void testThatTheTestsPassed() throws MavenInvocationException, IOException {
+        //we also check continuous testing
+        String executionDir = "projects/project-using-test-parameter-injection-processed";
+        testDir = initProject("projects/project-using-test-parameter-injection", executionDir);
+        runAndCheck();
+
+        ContinuousTestingMavenTestUtils testingTestUtils = new ContinuousTestingMavenTestUtils(getPort());
+        ContinuousTestingMavenTestUtils.TestStatus results = testingTestUtils.waitForNextCompletion();
+        // This is a bit brittle when we add tests, but failures are often so catastrophic they're not even reported as failures,
+        // so we need to check the pass count explicitly
+        Assertions.assertEquals(0, results.getTestsFailed());
+        Assertions.assertEquals(1, results.getTestsPassed());
+    }
+
+}
